# Backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pigeonchat API",
        "version": "1.0.0",
        "description": "API for pigeonchat application",
        "contact": {
            "name": "Vivek Maniyar",
            "email": "vivekmaniyar9601@gmail.com"
        }
    },
    "paths":{
...
        "/chats":{
            "post":{
                "tags":[
                    "Chats"
                ],
                "summary":"Create a new group chat",
                "description":"Create a new group chat",
                "operationId":"createChat",
                "requestBody": {
                    "description": "Chat to create",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/creategroup"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Chat created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Chat"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "get":{
                "tags":[
                    "Chats"
                ],
                "summary":"Get all chats",
                "description":"Get all chats",
                "operationId":"getChats",
                "responses": {
                    "200": {
                        "description": "Chats returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Chat"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chats/getchat":{
            "post":{
                "tags":[
                    "Chats"
                ],
                "summary":"Get a chat",
                "description":"Get a chat",
                "operationId":"getChat",
                "requestBody": {
                    "description": "Chat to get",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/getchat"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Chat returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Chat"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "New chat created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Chat"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chats/{chatId}":{
            "post":{
                "tags":[
                    "Chats"
                ],
                "summary":"Add users to a chat",
                "description":"Add users to a chat",
                "operationId":"addUsersToChat",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID of chat to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Users to add to chat",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/addUsersToChat"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Users added to chat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Chat"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete":{
                "tags":[
                    "Chats"
                ],
                "summary":"Delete group chat",
                "description":"Delete group chat",
                "operationId":"deleteChat",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID of chat to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Chat deleted"
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put":{
                "tags":[
                    "Chats"
                ],
                "summary":"Update group chat name",
                "description":"Update group chat name",
                "operationId":"updateChat",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID of chat to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Chat to update",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateChatName"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Chat name updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Chat"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chats/removeuser/{chatId}":{
            "delete":{
                "tags":[
                    "Chats"
                ],
                "summary":"Remove user from a chat",
                "description":"Remove user from a chat",
                "operationId":"removeUserFromChat",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID of chat to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User to remove from chat",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RemoveUser"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User removed from chat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Chat"
                                }
                            }
                        }
                    },
                    "500":{
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "401":{
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components":{
        "schemas":{
            "error":{
                "type":"object",
                "properties":{
                    "name":{
                        "type":"string"
                    },
                    "message":{
                        "type":"string"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties":{
                    "name": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "profileimage": {
                        "type": "string"
                    }
                }
            },
            "Message":{
                "type":"object",
                "properties":{
                    "message":{
                        "type":"string"
                    },
                    "sender":{
                        "$ref":"#/components/schemas/User"
                    },
                    "chat":{
                        "$ref":"#/components/schemas/Chat"
                    },
                    "readby":{
                        "type":"array",
                        "items":{
                            "$ref":"#/components/schemas/User"
                        }
                    }
                }
            },
            "Chat":{
                "type":"object",
                "properties":{
                    "chatName":{
                        "type":"string"
                    },
                    "isGroupChat":{
                        "type":"boolean"
                    },
                    "users":{
                        "type":"array",
                        "items":{
                            "$ref":"#/components/schemas/User"
                        }
                    },
                    "latestMessage":{
                        "$ref":"#/components/schemas/Message"
                    },
                    "groupadmin":{
                        "$ref":"#/components/schemas/User"
                    }
                }
            },
...
            "getchat":{
                "type":"object",
                "properties":{
                    "user":{
                        "type":"string"
                    }
                }
            },
...
            "creategroup":{
                "type":"object",
                "properties":{
                    "chatName":{
                        "type":"string"
                    },
                    "isGroupChat":{
                        "type":"boolean"
                    },
                    "users":{
                        "type":"array",
                        "items":{
                            "type":"string"
                        }
                    }
                }
            },
            "addUsersToChat":{
                "type":"object",
                "properties":{
                    "users":{
                        "type":"array",
                        "items":{
                            "type":"string"
                        }
                    }
                }
            },
...
            "UpdateChatName":{
                "type":"object",
                "properties":{
                    "chatName":{
                        "type":"string"
                    }
                }
            },
            "RemoveUser":{
                "type":"object",
                "properties":{
                    "user":{
                        "type":"string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    }
}